requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
python-jose==3.3.0
cryptography==41.0.7
PyJWT==2.8.1
httpx==0.25.2
python-dotenv==1.0.0